requirements.txt
playwright==1.42.0
playwright-stealth==1.0.6
python-dotenv
streamlit>=1.32.0
supabase>=2.0.0
plotly>=5.18.0
pandas>=2.0.0
supabase==2.3.4
openai==1.14.0
pyyaml==6.0.1
